+++ packages/data-types/src/lib/interfaces/pet-shop-adoptions.interface.ts
@@ -1,0 +1,9 @@
+export interface PetShopAdoptions{
+    npcData: {
+        npcId: string;
+        nickname: string;
+    }
+    price: number;
+    description: string;
+    iconName: string;
+}