+++ packages/guide/src/app/locations/components/pet-adoption-details/pet-adoption-details.component.ts
@@ -1,0 +1,27 @@
+import { Component, inject, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
+import { NPC, PetShopAdoptions } from "@ci/data-types";
+import { DatabaseService } from "../../../shared/services/database.service";
+import { UiIcon } from "../../../shared/enums/ui-icon.enum";
+
+@Component({
+    selector: 'app-pet-adoption-details',
+    templateUrl: './pet-adoption-details.component.html',
+})
+export class PetAdoptionDetailsComponent implements OnInit, OnChanges {
+
+    @Input({required: true}) adoption!: PetShopAdoptions;
+
+    npc?: NPC;
+    protected readonly UiIcon = UiIcon;
+    private _database: DatabaseService = inject(DatabaseService)
+
+    ngOnInit(): void {
+        this.npc = this._database.getNPCs().find(npc => npc.key === this.adoption.npcData.npcId);
+    }
+
+    ngOnChanges(changes: SimpleChanges): void {
+        if (changes['adoption'] && !changes['adoption'].isFirstChange()) {
+            this.npc = this._database.getNPCs().find(npc => npc.key === changes['adoption'].currentValue.npcData.npcId);
+        }
+    }
+}